% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_parameters.R
\name{backtest_parameters}
\alias{backtest_parameters}
\title{Factory for fixing dataset and backtesting strategy with different parameters
Useful for parameter optimisation}
\usage{
backtest_parameters(
  X,
  signal_maker = make_macd_signals,
  backtester = backtest_signals,
  simplistic = TRUE
)
}
\arguments{
\item{X}{xts}

\item{signal_maker}{function}

\item{backtester}{function}

\item{simplistic}{logical}
}
\value{
function
}
\description{
Factory for fixing dataset and backtesting strategy with different parameters
Useful for parameter optimisation
}
