% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_period.R
\name{backtest_period}
\alias{backtest_period}
\title{Factory for fixing parameters and backtesting strategy on different periods}
\usage{
backtest_period(
  signal_maker = make_macd_signals,
  maker_parameters,
  backtester = backtest_signals,
  tester_parameters,
  simplistic = TRUE
)
}
\arguments{
\item{signal_maker}{function}

\item{maker_parameters}{list}

\item{backtester}{function}

\item{tester_parameters}{list}

\item{simplistic}{logical}
}
\value{
function
}
\description{
Factory for fixing parameters and backtesting strategy on different periods
}
