% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_signals.R
\name{backtest_signals}
\alias{backtest_signals}
\title{Basic fast backtester}
\usage{
backtest_signals(
  X,
  signals_column = "signal_inter",
  simple_result = FALSE,
  starting_capital = 1000,
  starting_stock = 0,
  comission = 7e-04
)
}
\arguments{
\item{X}{xts}

\item{signals_column}{character}

\item{simple_result}{logical}

\item{starting_capital}{integer}

\item{starting_stock}{integer}

\item{comission}{double}
}
\value{
double
}
\description{
Basic fast backtester
}
\examples{
signals = na.omit(make_macd_signals(AFLT2014, nFast = 50, nSlow = 200, nSig = 9))
data = na.omit(cbind(AFLT2014, signals))
backtest_signals(data)
}
