% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_signals_slow.R
\name{backtest_signals_slow}
\alias{backtest_signals_slow}
\title{Slow backtester with loop. Use it for debug purposes}
\usage{
backtest_signals_slow(
  X,
  signals_column = "signal_inter",
  verbose = 0,
  simple_result = FALSE,
  starting_capital = 1000,
  starting_stock = 0,
  comission = 7e-04
)
}
\arguments{
\item{X}{xts}

\item{signals_column}{character}

\item{verbose}{integer}

\item{simple_result}{logical}

\item{starting_capital}{integer}

\item{starting_stock}{integer}

\item{comission}{double}
}
\value{
double
}
\description{
Slow backtester with loop. Use it for debug purposes
}
