% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly.R
\name{butterfly}
\alias{butterfly}
\title{Butterfly: compare previous data in continuously updated timeseries}
\usage{
butterfly(df_current, df_previous, datetime_variable)
}
\arguments{
\item{df_current}{data.frame, most recent dataset n.}

\item{df_previous}{data.frame, the previoust dataset, ie n - 1.}

\item{datetime_variable}{string, which unique ID to use to join x and y. Usually "time".}
}
\description{
This function compares two timeseries objects where we expect previous values
to be the same.
}
\details{
This function is intended to aid in the QA/QC of continually updating
timeseries data where we expect new values, but want to ensure previous data
remains unchanged.
}
