% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_catch.R
\name{butterfly_catch}
\alias{butterfly_catch}
\title{Butterfly catch: return dataframe containing only rows that have changed}
\usage{
butterfly_catch(df_current, df_previous, datetime_variable)
}
\arguments{
\item{df_current}{data.frame, most recent dataset n.}

\item{df_previous}{data.frame, the previous dataset, ie n - 1.}

\item{datetime_variable}{string, which unique ID to use to join df_current and df_previous. Usually a "datetime" variable.}
}
\description{
This function matches two dataframe objects by their unique identifier
(usually "time" or "datetime in a timeseries), and returns a new dataframe
which contains only rows that have changed compared to previous data. It will
not return any new rows.
}
