% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_release.R
\name{butterfly_release}
\alias{butterfly_release}
\title{Butterfly release: return current dataframe without changed old rows}
\usage{
butterfly_release(df_current, df_previous, datetime_variable)
}
\arguments{
\item{df_current}{data.frame, most recent dataset n.}

\item{df_previous}{data.frame, the previous dataset, ie n - 1.}

\item{datetime_variable}{string, which unique ID to use to join df_current and df_previous. Usually a "datetime" variable.}
}
\description{
This function matches two dataframe objects by their unique identifier
(usually "time" or "datetime in a timeseries), and returns a new dataframe
which contains the new rows (if present) but matched rows which contain
changes from previous data will be dropped.
}
