% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterflycount}
\alias{butterflycount}
\title{Butterfly count dummy data}
\format{
\subsection{\code{butterflycount}}{

A list with 4 dataframes (january, february, march, april) containing
3 columns, and 3 + n_month rows:
\describe{
\item{time}{The date on which the count took place, in yyyy-mm-dd format}
\item{count}{Number of fictional butterflies counted}
\item{species}{Butterfly species name, only appears in april}
...
}
}
}
\usage{
butterflycount
}
\description{
A completely fictional dataset of monthly butterfly counts
}
\keyword{datasets}
