% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{butterflymess}
\alias{butterflymess}
\title{Butterfly count messy data}
\format{
\subsection{\code{butterflymess}}{

A list with 5 dataframes (january, february, march, april, may) containing
3 columns, and 3 + n_month rows:
\describe{
\item{time}{The date on which the imaginary, and messy, count took place,
in yyyy-mm-dd format}
\item{count}{Number of fictional butterflies counted}
\item{species}{Butterfly species name, only appears in april}
...
}
}
}
\usage{
butterflymess
}
\description{
A version of butterflycount made messy using the messy package. This dataset
is only used for testing purposes
}
\keyword{datasets}
