% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forestprecipitation}
\alias{forestprecipitation}
\title{Forest precipitation dummy data}
\format{
\subsection{\code{butterflycount}}{

A list with 2 dataframes (january, february) containing 2 columns,
and 6 rows. February intentionally resets to 1970-01-01
\describe{
\item{time}{The date on which the imaginary rainfall was measured took
place, in yyyy-mm-dd format}
\item{rainfall_mm}{Rainfall in mm}
...
}
}
}
\usage{
forestprecipitation
}
\description{
A completely fictional dataset of daily precipitation
}
\keyword{datasets}
