% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contribs.R
\name{get_contribs}
\alias{get_contribs}
\title{Retrieve Contribution Data for a GitHub User and Chart it}
\usage{
get_contribs(user)
}
\arguments{
\item{user}{GitHub user (handle)}
}
\value{

}
\description{
Retrieve Contribution Data for a GitHub User and Chart it
}

