% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github.R
\name{get_github_chart}
\alias{get_github_chart}
\title{Chart GitHub User Activity}
\usage{
get_github_chart(user, addCommits = TRUE, showPlot = TRUE)
}
\arguments{
\item{user}{GitHub user (handle)}

\item{addCommits}{logical. Should the number of commits be overlayed?}

\item{showPlot}{logical. Should the plot be displayed? Otherwise 
\code{ggplot2} object is just returned.}
}
\value{
a list containing the ggplot object and data object:
  
  list(plot = <ggplot2 object>, data = <data.frame>)
}
\details{
As per the API 
  \url{https://help.github.com/articles/why-are-my-contributions-not-showing-up-on-my-profile/},
   returns a \code{data.frame} of columns containing data obtained from the 
  requested \code{user}'s contributions chart, typically found at
  \url{https://github.com/users/<USER>/contributions}.
  
  This function utilises both \code{\link[githubtools]{theme_github}} and 
  \code{\link[githubtools]{scale_fill_social}} and is subject to updates to 
  those functions.
}
\examples{
\dontrun{
library(butteRfly)
get_github_chart("jonocarroll") # Jonathan Carroll
get_github_chart("hadley")      # Hadley Wickham}
}

