% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stackoverflow.R
\name{get_stackoverflow}
\alias{get_stackoverflow}
\title{Retrieve Answers Data for a Stack Overflow User and Chart it}
\usage{
get_stackoverflow(user)
}
\arguments{
\item{user}{StackOverflow user id}
}
\value{
a \code{ggplot} object of the user's answers (and whether they were accepted)
}
\description{
Retrieve Answers Data for a Stack Overflow User and Chart it
}

