% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter.R
\name{get_twitter_chart}
\alias{get_twitter_chart}
\title{Chart Twitter Tweet Activity}
\usage{
get_twitter_chart(user = NULL, loadSaved = NULL, addNumbers = TRUE,
  showPlot = TRUE)
}
\arguments{
\item{user}{Twitter user (handle)}

\item{loadSaved}{if provided, a saved output from this function for which to
repeat evaluation}

\item{addNumbers}{logical. Should number of Tweets be overlayed?}

\item{showPlot}{logical. Should the plot be displayed? Otherwise
\code{ggplot2} object is just returned.}
}
\value{
a list containing the ggplot object and data object:
  
  list(plot = <ggplot2 object>, data = <data.frame>)
}
\description{
This function *does* require you to be authenticated. Follow
  directions from \code{?twitter_setup} (construct a \code{~/.twitter} file
  with OAuth credentials). The search is performed using
  \code{\link[twitteR]{userTimeline}} which has a rate limit maximum of 3200
  Tweets. Allowances are made for this liimt.
}
\details{
As per the API 
  \url{https://dev.twitter.com/rest/reference/get/statuses/user_timeline}, "[sic]
  Native retweets of other statuses by the user is included in this total, 
  regardless of whether include_rts is set to false when requesting this 
  resource." so this is unavoidable behaviour.
  
  This function utilises both \code{\link[githubtools]{theme_github}} and
  \code{\link[githubtools]{scale_fill_social}} and is subject to updates to
  those functions.
}
\examples{
\dontrun{
get_twitter_chart("carroll_jono")}
}

