% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_macd_signals.R
\name{make_macd_signals}
\alias{make_macd_signals}
\title{Making MACD signals}
\usage{
make_macd_signals(
  X,
  nFast = 12,
  nSlow = 26,
  nSig = 9,
  simple_result = FALSE,
  volume_threshold = 0,
  macd_threshold = 0
)
}
\arguments{
\item{X}{xts}

\item{nFast}{integer}

\item{nSlow}{integer}

\item{nSig}{integer}

\item{simple_result}{logical}

\item{volume_threshold}{double}

\item{macd_threshold}{double}
}
\value{
xts
}
\description{
Making MACD signals
}
\examples{
signals = make_macd_signals(AFLT2014, nFast = 50, nSlow = 200, nSig = 9)
}
