% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{twitter_setup}
\alias{twitter_setup}
\title{Perform Twitter OAuth setup.}
\usage{
twitter_setup(config_file = "~/.twitter", echo = FALSE)
}
\value{
NULL (used for side-effect)
}
\description{
#@param config_file location of the configuration file, default ~/.twitter
}
\details{
#@param echo if TRUE print the credentials read from the file to the console.
}
\examples{
\dontrun{
twitter_setup()
}
}

