% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly-package.R
\docType{package}
\name{butterfly-package}
\alias{butterfly}
\alias{butterfly-package}
\title{butterfly: QA/QC For Continually Updating Timeseries Data}
\description{
QA/QC of continually updating timeseries data where we expect new values, but want to ensure previous data remains unchanged.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomaszwagerman/butterfly}
  \item Report bugs at \url{https://github.com/thomaszwagerman/butterfly/issues}
}

}
\author{
\strong{Maintainer}: Thomas Zwagerman \email{thozwa@bas.ac.uk} (\href{https://orcid.org/0009-0003-3742-3234}{ORCID})

}
\keyword{internal}
