% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly-package.R
\docType{package}
\name{butterfly-package}
\alias{butterfly}
\alias{butterfly-package}
\title{butterfly: Verification for Continually Updating Time Series Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Verification of continually updating time series data where we expect new values, but want to ensure previous data remains unchanged. Data previously recorded could change for a number of reasons, such as discovery of an error in model code, a change in methodology or instrument recalibration. Monitoring data sources for these changes is not always possible. Other unnoticed changes could include a jump in time or measurement frequency, due to instrument failure or software updates. This package provides functionality that can be used as part of a data pipeline, to check and flag changes to previous data to prevent changes going unnoticed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/butterfly/}
  \item \url{https://github.com/ropensci/butterfly/}
  \item Report bugs at \url{https://github.com/ropensci/butterfly/issues}
}

}
\author{
\strong{Maintainer}: Thomas Zwagerman \email{thozwa@bas.ac.uk} (\href{https://orcid.org/0009-0003-3742-3234}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Quentin Read (Quentin reviewed the package (v. 1.1.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/676>) [reviewer]
}

}
\keyword{internal}
